Account for extra newlines in response body
The timeout=1 (1ms) parameter would some times trigger extra newlines to be
included in the response body. The use of `binary:split/2` would then
return different portions of the body depending on timing in the
cluster. This change adds a helper function to split out all newlines in
the response and then returns the last non-empty line.

This also removes introspection of the clustered update sequence since
this is an HTTP API behavior tests and those are defined as opaque
values.

COUCHDB-3415