Remove LRU from ddoc_cache
This is only a part of the necessary work to optimize the ddoc cache.
This will not work in production because the ets_lru max_age is not
implemented so once a ddoc is in the cache it won't be removed unless
everything stops using it (which for a busy database will be never
because of the match_newest code).