Generalize cache entry specific logic
There was a lot of guff going on with all of the various special cased
function clauses to separate out different behavior. This makes things a
lot more straight forward in terms of isolating the different logic.

The one slight downside is that there's a somewhat complex change to how
we manage evictions now that the ets table keys are opaque. Though on
the plus side this changed turned an ets table scan into a set of O(1)
khash lookups, so there's that at least.

This will still not work in production as we're not kicking out entries
when their existence changes. Patch three in this series will address
that last detail.