Refresh cache entries periodically
This patch restores the general runtime correctness of ddoc_cache. The
refresher is responsible for updating cache entries either when they
change or when they need to be removed. This clears up the previous
possible race condition between ddoc_cache_opener and ddoc_cache_lru
that could have left a stale entry in the cache. This same exact race
condition existed with ets_lru but the max_age of ets_lru would cover up
the issue if it ever happened (which would be fairly rare).

This change also finally addresses the major issue in ddoc_cache where
entries were forcibly evicted after max_age. This lead to situations
where we would have a thundering herd of clients trying to all insert
the same cache item just after it expired. In some extreme cases this
would lead to a backup of the ddoc_cache message queue that was
unrecoverable. Hopefully this change addresses that behavior.